+++ MOP/src/Common/BugReporter.cs
@@ -21,9 +21,9 @@ using System.Collections;
 using Ionic.Zip;
 using UnityEngine;
 using MSCLoader;
+using MSCLoader.Helper;
 
 using MOP.Helpers;
-using MSCLoader.Helper;
 
 namespace MOP.Common
 {