+++ MOP/src/FSM/Actions/CustomCreateHose.cs
@@ -14,36 +14,28 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.If not, see<http://www.gnu.org/licenses/>.
 
-using HutongGames.PlayMaker;
 using UnityEngine;
 
 using MOP.Items;
+using MOP.Managers;
 
 namespace MOP.FSM.Actions
 {
-    /// <summary>
-    /// This FsmAction role is to instantiate a gameobject.
-    /// </summary>
-    class CustomCreateObject : FsmStateAction
+    class CustomCreateHose : CustomCreateObject
     {
-        GameObject parent;
-        protected GameObject prefab;
-
-        protected GameObject newObject;
-
-        public CustomCreateObject(GameObject parent, GameObject prefab)
+        public CustomCreateHose(GameObject parent, GameObject prefab) : base(parent, prefab)
         {
-            this.parent = parent;
-            this.prefab = prefab;
+            GameObject newPrefab = GameObject.Instantiate(prefab);
+            newPrefab.name = newPrefab.name.Replace("(Clone)(Clone)", "(Clone)");
+            this.prefab = newPrefab;
+            Object.Destroy(this.prefab.GetComponent<ItemBehaviour>());
+            this.prefab.SetActive(false);
         }
 
         public override void OnEnter()
         {
-            newObject = GameObject.Instantiate(prefab);
-            newObject.transform.position = parent.transform.position;
-            newObject.name = newObject.name.Replace("(Clone)(Clone)", "(Clone)");
-            newObject.SetActive(true);
-            newObject.AddComponent<ItemBehaviour>();
+            base.OnEnter();
+            ItemsManager.Instance.SetCurrentRadiatorHose(newObject);
         }
     }
 }