+++ MOP/src/FSM/Actions/CustomSatsumaBumperAttach.cs
@@ -14,36 +14,18 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.If not, see<http://www.gnu.org/licenses/>.
 
-using HutongGames.PlayMaker;
-using UnityEngine;
-
-using MOP.Items;
+using MOP.Items.Cases;
 
 namespace MOP.FSM.Actions
 {
-    /// <summary>
-    /// This FsmAction role is to instantiate a gameobject.
-    /// </summary>
-    class CustomCreateObject : FsmStateAction
+    class CustomSatsumaBumperAttach : CustomSatsumaBumperDetach
     {
-        GameObject parent;
-        protected GameObject prefab;
-
-        protected GameObject newObject;
-
-        public CustomCreateObject(GameObject parent, GameObject prefab)
-        {
-            this.parent = parent;
-            this.prefab = prefab;
-        }
+        public CustomSatsumaBumperAttach(RearBumperBehaviour behaviour) : base(behaviour) { }
 
         public override void OnEnter()
         {
-            newObject = GameObject.Instantiate(prefab);
-            newObject.transform.position = parent.transform.position;
-            newObject.name = newObject.name.Replace("(Clone)(Clone)", "(Clone)");
-            newObject.SetActive(true);
-            newObject.AddComponent<ItemBehaviour>();
+            behaviour.OnAttach();
+            Finish();
         }
     }
 }