+++ MOP/src/FSM/Actions/CustomSatsumaBumperDetach.cs
@@ -21,7 +21,7 @@ namespace MOP.FSM.Actions
 {
     class CustomSatsumaBumperDetach : FsmStateAction
     {
-        RearBumperBehaviour behaviour;
+        protected RearBumperBehaviour behaviour;
 
         public CustomSatsumaBumperDetach(RearBumperBehaviour behaviour)
         {
@@ -31,22 +31,6 @@ namespace MOP.FSM.Actions
         public override void OnEnter()
         {
             behaviour.OnDetach();
-            Finish();
-        }
-    }
-
-    class CustomSatsumaBumperAttach : FsmStateAction
-    {
-        RearBumperBehaviour behaviour;
-
-        public CustomSatsumaBumperAttach(RearBumperBehaviour behaviour)
-        {
-            this.behaviour = behaviour;
-        }
-
-        public override void OnEnter()
-        {
-            behaviour.OnAttach();
             Finish();
         }
     }